Add a barebones integration with manga-ocr
Included is a proposed API to potentially support integration with
many different OCR solutions, using the "source" for the language.

I've also supposed that users will tend to prefer one language, and that
one language's OCR can be used for local source manga. Without adding a
way to assign a language to any given manga, this seems like a simple
solution for now---and doesn't close the door on adding that in the
future.

This abstraction also paves a path toward an Extension-style interface
for installing different OCR providers and removing the actual
implementation from being a core dependency. We would need to remove our
lazy use of `manga-ocr-android`'s models, of course...

---

All squashed commit messages below:

---

Show recognition progress + auto launch TranslationIntent on complete

This works *okay* but the model seems to perform worse when zoomed in,
which is annoying. As mentioned before we should probably pull out the
full bitmap and clip that---means we don't have to rely on disabling
hardware bitmaps, too.

Also need to try to get text box extraction working so we can tap
outside of text nicely

Add rough initial integration of TextDetector

Note that this doesn't actually work unless you're zoomed out (and maybe
not even then?) since our taps aren't looking at the "real" coords, but
zoomed coords

Convert to using a full bitmap for OCR, shared w/detection

This seems to work quite well! It should be less brittle than what we
tried to do before since we own the bitmap ourself, and it also means we
can re-enable hardware bitmaps

Restore hardware bitmap support

Pass more context to load bitmap more accurately

Important for getting coordinate mapping correct. It'd be nice if we
could get the bitmap from the view somehow but....

An alternative we could explore is figuring out how big the view thinks
its source is, and how big *we* think it is, and mapping additionally
there? This would allow us to load a smaller bitmap....

Not sure how well that idea plays with the cropBorders config (which was
important here)

Properly cancel an OCR session on dismiss

Okay... "properly." It's a bit of a kludge, but it works

Attempt to pre-emptively detect text on navigate

Relies on an update to manga-ocr-android that reduces duplicate regions.
Might be nice to also scan ahead to next pages...

Another nice feature here is we can release the bitmap when we know
we've fully scanned the page, potentially slightly easing memory
constraint concerns.

There's an edge case here where we navigate away from a page too
quickly, but that's what the LRU cache is for.

Upgrade manga-ocr-android with text box detection improvements

Go ahead and recycle the bitmap

Tweak BitmapConfig to try to reduce memory overhead

Disable eager OCR

This *feels* great in theory, but turns my phone into a hand warmer,
alas.

Upgrade ocr lib for huge speedups

We have to switch back to the default `ARGB_8888` bitmap format, at
least for now, but it's worth it.

This actually includes two speedups:

1. Converting a different text detector model to tflite, resulting in an
   order of magnitude speedup for Detector
2. Splitting the existing OCR model into a separate encoder and decoder
   step, so we can reuse the slow encoder outputs across all decoder
   iterations---resulting in an order of magnitude speedup for MangaOcr!

We've gone from ~3s to detect down to ~300ms, and ~100ms/token for OCR
down to ~10ms/token. This is MUCH nicer to use!

Upgrade manga-ocr again with more robust text box detection

Don't render Text on the progress modal if it's empty

Slightly cleaner to avoid the unnecessary padding

Propose an abstraction for multiple OCR providers; use source for lang

I've also supposed that users will tend to prefer one language, and that
one language's OCR can be used for local source manga. Without adding a
way to assign a language to any given manga, this seems like a simple
solution for now---and doesn't close the door on adding that in the
future.

This abstraction also paves a path toward an Extension-style interface
for installing different OCR providers and removing the actual
implementation from being a core dependency. We would need to remove our
lazy use of `manga-ocr-android`'s models, of course...

Fix: handle potential out-of-bounds detected text boxes

Fix release builds with updated dependency including proguard rules

Upgrade manga-ocr-android for improved accuracy on color text

Fix remaining build issues from rebase + upgrade deps