Show recognition progress + auto launch TranslationIntent on complete
This works *okay* but the model seems to perform worse when zoomed in,
which is annoying. As mentioned before we should probably pull out the
full bitmap and clip that---means we don't have to rely on disabling
hardware bitmaps, too.

Also need to try to get text box extraction working so we can tap
outside of text nicely