Pass more context to load bitmap more accurately
Important for getting coordinate mapping correct. It'd be nice if we
could get the bitmap from the view somehow but....

An alternative we could explore is figuring out how big the view thinks
its source is, and how big *we* think it is, and mapping additionally
there? This would allow us to load a smaller bitmap....

Not sure how well that idea plays with the cropBorders config (which was
important here)