Attempt to pre-emptively detect text on navigate
Relies on an update to manga-ocr-android that reduces duplicate regions.
Might be nice to also scan ahead to next pages...

Another nice feature here is we can release the bitmap when we know
we've fully scanned the page, potentially slightly easing memory
constraint concerns.

There's an edge case here where we navigate away from a page too
quickly, but that's what the LRU cache is for.