Propose an abstraction for multiple OCR providers; use source for lang
I've also supposed that users will tend to prefer one language, and that
one language's OCR can be used for local source manga. Without adding a
way to assign a language to any given manga, this seems like a simple
solution for now---and doesn't close the door on adding that in the
future.

This abstraction also paves a path toward an Extension-style interface
for installing different OCR providers and removing the actual
implementation from being a core dependency. We would need to remove our
lazy use of `manga-ocr-android`'s models, of course...